+++ src/inspect_ai/scorer/__init__.py
@@ -24,6 +24,7 @@ from ._metric import (
 from ._metrics.accuracy import accuracy
 from ._metrics.categorical import categorical, frequency
 from ._metrics.grouped import grouped
+from ._metrics.krippendorff import krippendorff_alpha
 from ._metrics.mean import mean
 from ._metrics.perplexity import perplexity_per_seq, perplexity_per_token
 from ._metrics.std import bootstrap_stderr, std, stderr, var
@@ -35,6 +36,7 @@ from ._reducer import (
     ScoreReducer,
     ScoreReducers,
     at_least,
+    collect,
     max_score,
     mean_score,
     median_score,
@@ -72,11 +74,13 @@ __all__ = [
     "bootstrap_stderr",
     "categorical",
     "choice",
+    "collect",
     "exact",
     "f1",
     "frequency",
     "grouped",
     "includes",
+    "krippendorff_alpha",
     "match",
     "math",
     "max_score",